refactor provider passport list query service structure
