+++ backend/src/main/java/com/alligator/market/backend/provider/config/ProviderFeatureWiringConfig.java
@@ -1,7 +1,7 @@
 package com.alligator.market.backend.provider.config;
 
 import com.alligator.market.backend.provider.config.adapter.moex.iss.MoexIssProviderConfig;
-import com.alligator.market.backend.provider.config.application.passport.catalog.PassportCatalogServiceWiringConfig;
+import com.alligator.market.backend.provider.config.application.passport.query.list.PassportListServiceWiringConfig;
 import com.alligator.market.backend.provider.config.readmodel.passport.projection.startup.ProviderPassportProjectionStartupRunnerWiringConfig;
 import com.alligator.market.backend.provider.config.registry.ProviderRegistryWiringConfig;
 import org.springframework.context.annotation.Configuration;
@@ -13,7 +13,7 @@ import org.springframework.context.annotation.Import;
 @Configuration(proxyBeanMethods = false)
 @Import({
         ProviderRegistryWiringConfig.class,
-        PassportCatalogServiceWiringConfig.class,
+        PassportListServiceWiringConfig.class,
         ProviderPassportProjectionStartupRunnerWiringConfig.class,
         MoexIssProviderConfig.class
 })
